Revert incorrect 'fixes' and merge existing tests
Removes the incorrect test I added and undoes the fix that came with it.
Puts the existing tests into the new classes under the Input ns
Also fixes namespace to be consistent with other tests.
