+++ Test/Framework/Input/JoystickTest.cs
@@ -5,7 +5,7 @@
 using Microsoft.Xna.Framework.Input;
 using NUnit.Framework;
 
-namespace MonoGame.Tests
+namespace MonoGame.Tests.Input
 {
     public class JoystickTest
     {
